Changes after CR #1
