+++ tests/contrib/operators/test_cloudformation.py
@@ -21,10 +21,52 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from airflow import DAG
-from airflow.contrib.operators.cloudformation_delete_stack_operator import CloudFormationDeleteStackOperator
+from airflow.contrib.operators.cloudformation import CloudFormationCreateStackOperator, \
+    CloudFormationDeleteStackOperator
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2019, 1, 1)
+
+
+class TestCloudFormationCreateStackOperator(unittest.TestCase):
+
+    def setUp(self):
+        self.args = {
+            'owner': 'airflow',
+            'start_date': DEFAULT_DATE
+        }
+
+        # Mock out the cloudformation_client (moto fails with an exception).
+        self.cloudformation_client_mock = MagicMock()
+
+        # Mock out the emr_client creator
+        cloudformation_session_mock = MagicMock()
+        cloudformation_session_mock.client.return_value = self.cloudformation_client_mock
+        self.boto3_session_mock = MagicMock(return_value=cloudformation_session_mock)
+
+        self.mock_context = MagicMock()
+
+    def test_create_stack(self):
+        stack_name = "myStack"
+        timeout = 15
+        template_body = "My stack body"
+
+        operator = CloudFormationCreateStackOperator(
+            task_id='test_task',
+            params={
+                'StackName': stack_name,
+                'TimeoutInMinutes': timeout,
+                'TemplateBody': template_body
+            },
+            dag=DAG('test_dag_id', default_args=self.args),
+        )
+
+        with patch('boto3.session.Session', self.boto3_session_mock):
+            operator.execute(self.mock_context)
+
+        self.cloudformation_client_mock.create_stack.assert_any_call(StackName=stack_name,
+                                                                     TemplateBody=template_body,
+                                                                     TimeoutInMinutes=timeout)
 
 
 class TestCloudFormationDeleteStackOperator(unittest.TestCase):
@@ -46,11 +88,11 @@ class TestCloudFormationDeleteStackOperator(unittest.TestCase):
         self.mock_context = MagicMock()
 
     def test_delete_stack(self):
-        stack_name = "myStack"
+        stack_name = "myStackToBeDeleted"
 
         operator = CloudFormationDeleteStackOperator(
             task_id='test_task',
-            StackName=stack_name,
+            params={'StackName': stack_name},
             dag=DAG('test_dag_id', default_args=self.args),
         )
